fix test for successful comment deletion, add unsuccessful case #107

+++ collection/test_views.py
@@ -8,9 +8,19 @@ class TestCollectionViews(TestCase):
 
     def setUp(self):
         self.user = User.objects.create_superuser(
-            username="myUsername",
-            password="myPassword",
+            username="superUserName",
+            password="superUserPassword",
             email="test@test.com"
+        )
+        self.user1 = User.objects.create_user(
+            username="user1Name",
+            password="user1Password",
+            email="user1@test.com"
+        )
+        self.user2 = User.objects.create_user(
+            username="user2Name",
+            password="user2Password",
+            email="user2@test.com"
         )
         self.recipe = Recipe(title="Recipe title", 
                          author=self.user,
@@ -23,7 +33,7 @@ class TestCollectionViews(TestCase):
         self.recipe.save()
         self.comment = Comment(id=1,
                                 recipe=self.recipe,
-                                author=self.user,
+                                author=self.user1,
                                 body="Comment for deletion")
         self.comment.save()
 
@@ -39,7 +49,7 @@ class TestCollectionViews(TestCase):
     def test_successful_comment_submission(self):
         """Test for posting a comment on a recipe"""
         self.client.login(
-            username="myUsername", password="myPassword")
+            username="superUserName", password="superUserPassword")
         recipe_data = {
             'body': 'This is a test comment.'
         }
@@ -55,12 +65,27 @@ class TestCollectionViews(TestCase):
     def test_successful_comment_deletion(self):
         """Test for successfully deleting a comment"""
         self.client.login(
-            username="myUsername", password="myPassword")
-        self.comment.delete()
+            username="user1Name", password="user1Password")
+        if self.comment.author == self.user1:
+            self.comment.delete()
         response = self.client.get(reverse(
             'recipe_detail', args=['recipe-title']))
         self.assertEqual(response.status_code, 200)
         self.assertNotIn(
+            b'Comment for deletion',
+            response.content
+        )
+
+    def test_unsuccessful_comment_deletion(self):
+        """Test for failing to delete a comment belonging to another user"""
+        self.client.login(
+            username="user2Name", password="user2password")
+        if self.comment.author == self.user1:
+            self.comment.delete()
+        response = self.client.get(reverse(
+            'recipe_detail', args=['recipe-title']))
+        self.assertEqual(response.status_code, 200)
+        self.assertIn(
             b'Comment for deletion',
             response.content
         )
